request #11624 Projects that are pending for site admin validation trigger wrong email
How to test:
- for project approval
- create a new project
- without this patch site administrator receive an email saying that project
  was automatically approved (it's not)
- with this patch site admin receive an email saying that there is a project
  to approve (or not)

+++ src/common/project/Project.class.php
@@ -502,7 +502,7 @@ class Project extends Group implements PFO_Project {
     }
 
     public function projectsMustBeApprovedByAdmin() {
-        return ForgeConfig::get(\ProjectManager::CONFIG_PROJECT_APPROVAL, 1) === 1;
+        return (int) ForgeConfig::get(\ProjectManager::CONFIG_PROJECT_APPROVAL, 1) === 1;
     }
 
     /**
